+++ src/cloregram/callbacks.clj
@@ -1,5 +1,5 @@
 (ns cloregram.callbacks
-  (:require [dialog.logger :as log]
+  (:require [taoensso.timbre :as log]
             [cloregram.db :as db]
             [cloregram.utils :as utl]
             [datomic.api :as d]
@@ -10,26 +10,32 @@
   ([user ^clojure.lang.Symbol f args]
    (create (java.util.UUID/randomUUID) user f args))
   ([uuid user ^clojure.lang.Symbol f args]
-   (let [args (pr-str args)]
+   (let [args (or args {})]
      (d/transact (db/conn) [{:callback/uuid uuid
                              :callback/function f
-                             :callback/args args
+                             :callback/args (prn-str args)
                              :callback/user [:user/id (:user/id user)]}])
-     (log/debugf "Created Callback %s for User %s: %s %s" uuid user f args)
+     (log/debug "Created Callback" {:callback-uuid uuid
+                                    :callback-function f
+                                    :callback-arguments args
+                                    :user user})
      uuid)))
 
 (defn- load-callback
   [user uuid]
   (let [callback (d/pull (db/db) '[* {:callback/user [*]}] [:callback/uuid uuid])]
-    (log/debug (format "Loaded callback %s: %s" uuid callback))
+    (log/debug "Loaded callback" {:callback-uuid uuid
+                                  :callback-data callback})
     (when (not= (:id user) (-> callback :callback/user :id))
-      (throw (ex-info "Wrong User in loaded Callback!" {:user user :callback callback})))
+      (throw (ex-info "Wrong User attempt to load Callback!" {:user user :callback-data callback})))
     callback))
 
 (defn call
   [user ^java.util.UUID uuid]
   (let [callback (load-callback user uuid)
         func (:callback/function callback)
-        args (-> callback :callback/args edn/read-string)]
-    (log/debugf "Calling %s with args %s and user %s" func args user)
-    ((utl/resolver func) (assoc args :user user))))
+        args (or (-> callback :callback/args edn/read-string (assoc :user user)) {})]
+    (log/debug "Calling Callback function" {:callback-function func
+                                            :callback-arguments args
+                                            :user user})
+    ((utl/resolver func) args)))