+++ src/cloregram/db.clj
@@ -3,35 +3,35 @@
             [cloregram.system.state :refer [system]]
             [cloregram.utils :as utl]
             [cloregram.schema :refer [schema]]
-            [dialog.logger :as log]))
+            [taoensso.timbre :as log]))
 
 (defn conn
-  "Returns database connection"
+  "Returns Datiomic database connection"
   []
   (:db/connection @system))
 
 (defn db
-  "Returns current database state"
+  "Returns current Datomic database state"
   []
   (let [db (d/db (conn))]
-    (log/debug "Received current database state:" db)
+    (log/debug "Received current Datomic database state")
     db))
 
-(defn- read-user-schema
+(defn- read-project-schema
   []
   (utl/read-resource-dir "schema"))
 
 (defn update-schema
-  "Updates schema with new entities data. Not removing unactual."
+  "Updates Datomic database schema with new entities data. Not removing unactual."
   []
-  (log/info "Updating schema...")
-  (let [user-schema (read-user-schema)
-        full-schema (concat schema user-schema)]
-    (log/debug "Schema:" full-schema)
+  (log/info "Updating Datomic database schema...")
+  (let [project-schema (read-project-schema)
+        full-schema (concat schema project-schema)]
+    (log/debug "Datomic database schema loaded" {:schema full-schema})
     (let [f (d/transact (conn) full-schema)]
-      (log/info "Schema successfully updated with" (count (:tx-data @f)) "Datoms"))))
+      (log/debug "Datomic database schema successfully updated" {:datoms (:tx-data @f)}))))
 
-(defn- read-user-data
+(defn- read-project-data
   []
   (utl/read-resource-dir "data"))
 
@@ -39,7 +39,7 @@
   "Loads initial data. Re-write if something exists"
   []
   (log/info "Updating data...")
-  (let [user-data (read-user-data)]
-    (log/debug "Data:" user-data)
+  (let [user-data (read-project-data)]
+    (log/debug "Project data loaded" {:data user-data})
     (let [f (d/transact (conn) user-data)]
-      (log/info "Data successfully updated with" (count (:tx-data @f)) "Datoms"))))
+      (log/debug "Project data successfully uploaded" {:datoms (:tx-data @f)}))))