+++ src/cloregram/test/infrastructure/users.clj
@@ -1,5 +1,5 @@
 (ns cloregram.test.infrastructure.users
-  (:require [dialog.logger :as log]
+  (:require [taoensso.timbre :as log]
             [cloregram.test.infrastructure.state :as state]
             [cloregram.utils :as utl]))
 
@@ -20,16 +20,17 @@
               :messages (sorted-map)
               :waiting-for-response? false}]
     (swap! state/users #(assoc % uid user))
-    (log/debugf "Added user @%s. Total users count: %d" (name uid) (count @state/users))))
+    (log/info "Added virtuial user" {:virtual-user (uid @state/users)
+                                     :virtual-users-count (count @state/users)})))
 
 (defn- get-user-by-
   [key value]
-  (log/debugf "Getting User by %s -> %s" key value)
+  (log/debug "Getting virtual user..." {key value})
   (let [user (->> @state/users
                   (filter (fn [[k v]] (= value (key v))))
                   (first)
                   (val))]
-    (log/debug "Got User" user)
+    (log/debug "Got User" {:virtual-user user})
     user))
 
 (defn get-user-by-uid
@@ -37,47 +38,50 @@
   [uid] (get-user-by- :username (name uid)))
 
 (defn- main-message#
-  [uid]
-  (let [user (uid @state/users)
-        msgs (:messages user)
+  [user]
+  (let [msgs (:messages user)
         msg-id (:main-msg-id user)]
-    (msgs msg-id)))
+    (get msgs msg-id)))
+
+(defn- process-temp-messages
+  [f user]
+  (let [ msg-id (:main-msg-id user)
+        msgs (-> user :messages (dissoc msg-id))
+        msgs# (filter #(not (contains? (second %) :silent)) msgs)]
+    (f msgs#)))
 
 (defn- count-temp-messages#
-  [uid]
-  (let [user (uid @state/users)
-        msg-id (:main-msg-id user)
-        msgs (-> user :messages (dissoc msg-id))
-        msgs# (filter #(not (contains? (second %) :silent)) msgs)]
-    (count msgs#)))
+  [user]
+  (process-temp-messages count user))
 
 (defn- last-temp-message#
-  [uid]
-  (let [user (uid @state/users)
-        msg-id (:main-msg-id user)
-        msgs (-> user :messages (dissoc msg-id))
-        msgs# (filter #(not (contains? (second %) :silent)) msgs)]
-    (some-> msgs# last val)))
+  [user]
+  (process-temp-messages #(some-> % last val) user))
 
 (defn- get-response-or-current
   [uid f s timeout]
-  (log/debugf "Getting %s for User %s (timeout: %d)" s uid timeout)
-  (let [interval 100]
-    (loop [t timeout]
-      (cond (not (:waiting-for-response? (uid @state/users)))
-            (let [resp (f uid)]
-              (log/debugf "User %s got %s: %s" uid s resp)
+  (let [user (uid @state/users)
+        interval 100]
+    (log/info (str "Getting " s) {:waiting-for-response? (:waiting-for-response? user)
+                                  :virtual-user user
+                                  :timeout timeout})
+    (loop [t timeout
+           u user]
+      (cond (not (:waiting-for-response? u))
+            (let [resp (f u)]
+              (log/info (str "Got " s) {:virtual-user u
+                                        s resp})
               resp)
 
             (= 0 t) (throw (ex-info (format "No %s!" s) {:timeout timeout}))
 
             :else (do (Thread/sleep interval)
-                      (recur (- t interval)))))))
+                      (recur (- t interval) (uid @state/users)))))))
 
 (defn main-message
-  
+
   "Returns main message of virtual user with key `uid` from test infrastructure virtual users state storage. If virtual user is not awaiting response, then message structure or nil is returned immidiately. If virtual user is awaiting response, then function will wait until response or `timeout` milliseconds (default 2000). If awaiting of response not finished while timeout passed, Exception will be throwed."
-  
+
   {:added "0.8"}
 
   ([uid] (main-message uid 2000))